+++ astropop/pipelines/common/stages.py
@@ -2,6 +2,7 @@
 
 from ..core import Stage
 
+# TODO: Continue de design
 
 __all__ = ['ImageProcessing', 'SourceExtraction', 'AperturePhotometry',
            'PSFPhotometry', 'AstrometryCalib', 'SavePhotometryFITS',
@@ -10,8 +11,22 @@ __all__ = ['ImageProcessing', 'SourceExtraction', 'AperturePhotometry',
 
 
 class ImageProcessing(Stage):
+    """Basic image processing for single CCD products.
+
+    Product Needs:
+        Capabilities:
+            raw_filenames = list of raw science files to process
+            calib_filenames = list of calibration frames filenames
+        Instrument:
+            All singleccd instrument functions.
+    """
+    _requested_capabilities = ['raw_filenames', 'calib_filenames']
+
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
+
+    def run(self, product, config):
+        return
 
 
 class SourceExtraction(Stage):
@@ -39,6 +54,11 @@ class PhotometryCalib(Stage):
         super().__init__(*args, **kwargs)
 
 
+class PolarimetryCompute(Stage):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+
 class SavePhotometryFITS(Stage):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -62,5 +82,3 @@ class SaveImageASDF(Stage):
 class MultiImageProcessing(Stage):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
-
-# TODO: Continue de design