+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Important Changes:
 
+version 0.4.0:
+  - New data based processing workflow for generic pipelines
+  - Now that CCDData is an astropy class, all the image workflow
+    was ported to CCDData (again)
+  - Better logging for all processing functions
+
 version 0.2.2-dev:
   - Alpha for map polarimetry ploting
   - More direct way to calculate polarimetry angle (no more pccdpack way)